LoaderTool: Add support for loading (ACPI case) Kernel without SD card.
Story is that previously we had implemented shell command "bl" in LoaderTool for loading kernel which is packed into uefi as a payload. And it can only support to jump to start address. It works fine in case of device tree.

However, when we want to start kernel with ACPI support, we would have to regard linux kernel as a UEFI application (get entry point, pass EFI_HANDLE and SystemTable into it as well).
Then, we could meet this requirement by using a formatted SD card (with the help of uefi shell, execute "./xxx" to start that application).

Now, for the case of SD controller is unavailable, we add one more option 'efi_stub' for command "bl" to support loading kernel with ACPI enabled.

When executing command like "bl efi_stub param1 param2 ...", it will try to find the entry point and do all things as the case of using shell to execute an UEFI application.
(param1 param2 ... are kernel params)

That's to say, now we have 3 ways to run "bl" as following:

1. "bl jump": Jump to kernel image (as a FV in edk2-platforms/Silicon/Bosc/KunMingHu/Loader/Image) in uefi.

2. "bl efi_stub param1 param2 ...": Try to load kernel with specified kernel params, such as:
   "bl efi_stub console=ttyS0,115200 earlycon=sbi acpi=force"

3. help: show above usage.

Signed-off-by: Guo Yaxing <guoyaxing@bosc.ac.cn>